Move metbabase/utils/click-behavior into metabase/dashboard
