+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/ClickBehaviorSidebarContent.tsx
@@ -2,7 +2,7 @@ import { getIn } from "icepick";
 import { useMemo } from "react";
 
 import { getDashcardData } from "metabase/dashboard/selectors";
-import { isTableDisplay } from "metabase/utils/click-behavior";
+import { isTableDisplay } from "metabase/dashboard/utils";
 import { useSelector } from "metabase/utils/redux";
 import type { UiParameter } from "metabase-lib/v1/parameters/types";
 import type {
