+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/ClickBehaviorSidebarHeader/ClickBehaviorSidebarHeader.tsx
@@ -1,7 +1,7 @@
 import { jt, t } from "ttag";
 
+import { isTableDisplay } from "metabase/dashboard/utils";
 import { Box, Icon } from "metabase/ui";
-import { isTableDisplay } from "metabase/utils/click-behavior";
 import type { DashboardCard, DatasetColumn } from "metabase-types/api";
 
 import S from "../ClickBehaviorSidebar.module.css";
