+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/LinkOptions/CustomURLPicker.tsx
@@ -4,6 +4,7 @@ import { useCallback, useState } from "react";
 import { t } from "ttag";
 
 import CS from "metabase/css/core/index.css";
+import { isTableDisplay } from "metabase/dashboard/utils";
 import {
   Button,
   Flex,
@@ -12,7 +13,6 @@ import {
   Text,
   TextInputBlurChange,
 } from "metabase/ui";
-import { isTableDisplay } from "metabase/utils/click-behavior";
 import type { UiParameter } from "metabase-lib/v1/parameters/types";
 import { clickBehaviorIsValid } from "metabase-lib/v1/parameters/utils/click-behavior";
 import type {
