+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/LinkOptions/LinkOptions.tsx
@@ -3,8 +3,8 @@ import { useCallback } from "react";
 import { t } from "ttag";
 
 import CS from "metabase/css/core/index.css";
+import { isTableDisplay } from "metabase/dashboard/utils";
 import type { IconName } from "metabase/ui";
-import { isTableDisplay } from "metabase/utils/click-behavior";
 import type { UiParameter } from "metabase-lib/v1/parameters/types";
 import type {
   ArbitraryCustomDestinationClickBehavior,
