+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/TableClickBehaviorView/TableClickBehaviorView.tsx
@@ -4,13 +4,14 @@ import { t } from "ttag";
 import _ from "underscore";
 
 import CS from "metabase/css/core/index.css";
-import { hasActionsMenu } from "metabase/utils/click-behavior";
 import type {
   ClickBehavior,
   ClickBehaviorType,
   DashboardCard,
   DatasetColumn,
 } from "metabase-types/api";
+
+import { hasActionsMenu } from "../../../utils";
 
 import { Column } from "./Column";
