+++ frontend/src/metabase/dashboard/hooks/use-click-behavior-data.ts
@@ -11,12 +11,18 @@ import {
   getParameterValuesBySlugMap,
   getParameters,
 } from "metabase/dashboard/selectors";
+import { Dashboards } from "metabase/entities/dashboards";
+import { Questions } from "metabase/entities/questions";
 import { getUserAttributes } from "metabase/selectors/user";
-import { getLinkTargets } from "metabase/utils/click-behavior";
 import { useStore } from "metabase/utils/redux";
 import type { ClickObject } from "metabase/visualizations/types";
 import Question from "metabase-lib/v1/Question";
-import type { DashCardId } from "metabase-types/api";
+import type {
+  ClickBehavior,
+  DashCardId,
+  EntityCustomDestinationClickBehavior,
+  VisualizationSettings,
+} from "metabase-types/api";
 import type { State } from "metabase-types/store";
 
 type EntityObject = {
@@ -123,3 +129,42 @@ export const useClickBehaviorData = ({
 
   return { getExtraDataForClick };
 };
+
+export function getLinkTargets(settings?: VisualizationSettings) {
+  const { click_behavior, column_settings = {} } = settings || {};
+  return [
+    click_behavior,
+    ...Object.values(column_settings).map(
+      (settings) => settings.click_behavior,
+    ),
+  ]
+    .filter(hasLinkedQuestionOrDashboard)
+    .map(mapLinkedEntityToEntityQuery);
+}
+
+function hasLinkedQuestionOrDashboard({
+  type,
+  linkType,
+}: {
+  type?: ClickBehavior["type"];
+  linkType?: EntityCustomDestinationClickBehavior["linkType"];
+} = {}) {
+  if (type === "link") {
+    return linkType === "question" || linkType === "dashboard";
+  }
+  return false;
+}
+
+function mapLinkedEntityToEntityQuery({
+  linkType,
+  targetId,
+}: {
+  linkType: EntityCustomDestinationClickBehavior["linkType"];
+  targetId: EntityCustomDestinationClickBehavior["targetId"];
+}) {
+  return {
+    entity: linkType === "question" ? Questions : Dashboards,
+    entityType: linkType,
+    entityId: targetId,
+  };
+}
